Extract all helper functions into `client/utils`
