+++ client/src/utils.rs
@@ -1,0 +1,289 @@
+use std::collections::HashMap;
+use std::fs::{self, OpenOptions};
+use std::io::Write;
+use std::os::unix::fs::OpenOptionsExt;
+use std::path::{Path, PathBuf};
+
+use anyhow::{Context, Result, anyhow};
+use bitcoin::address::NetworkUnchecked;
+use bitcoin::key::Keypair;
+use bitcoin::{Address, Network, PublicKey, Txid};
+
+/// Parse and validate a Bitcoin address for the specified network.
+///
+/// # Arguments
+///
+/// * `addr` - The address string to parse.
+/// * `network` - The expected Bitcoin network.
+///
+/// # Returns
+///
+/// The parsed and validated address.
+///
+/// # Errors
+///
+/// Returns an error if:
+/// - The address format is invalid
+/// - The address is for a different network
+///
+/// # Example
+///
+/// ```ignore
+/// let address = parse_btc_address(
+///     "bcrt1qw508d6qejxtdg4y5r3zarvary0c5xw7kygt080",
+///     Network::Regtest
+/// )?;
+/// ```
+pub fn parse_btc_address(addr: &str, network: Network) -> Result<Address> {
+    addr.parse::<Address<NetworkUnchecked>>()
+        .context("Invalid Bitcoin address format")
+        .and_then(|unchecked| {
+            unchecked
+                .require_network(network)
+                .map_err(|_| anyhow!("Address is not valid for {} network", network))
+        })
+}
+
+/// Parse a network string into a Bitcoin network type.
+///
+/// Supports common aliases for each network type.
+///
+/// # Supported Values
+///
+/// | Network  | Accepted Values                    |
+/// |----------|-------------------------------------|
+/// | Mainnet  | `mainnet`, `main`, `bitcoin`        |
+/// | Testnet  | `testnet`, `test`, `testnet3`       |
+/// | Signet   | `signet`                            |
+/// | Regtest  | `regtest`, `reg`                    |
+///
+/// # Arguments
+///
+/// * `network` - The network string to parse.
+///
+/// # Returns
+///
+/// The corresponding `Network` enum value.
+///
+/// # Errors
+///
+/// Returns an error if the network string is not recognized.
+pub fn parse_btc_network(network: &str) -> Result<Network> {
+    match network.to_lowercase().as_str() {
+        "mainnet" | "main" | "bitcoin" => Ok(Network::Bitcoin),
+        "testnet" | "test" | "testnet3" => Ok(Network::Testnet),
+        "signet" => Ok(Network::Signet),
+        "regtest" | "reg" => Ok(Network::Regtest),
+        _ => Err(anyhow!(
+            "Invalid network '{}'. Supported: mainnet, testnet, signet, regtest",
+            network
+        )),
+    }
+}
+
+/// Parse and validate a Bitcoin keypair from WIF format.
+///
+/// # Arguments
+///
+/// * `key` - The private key in WIF (Wallet Import Format) or hex.
+/// * `role` - Description of the key's purpose (for error messages).
+///
+/// # Returns
+///
+/// The parsed keypair.
+///
+/// # Errors
+///
+/// Returns an error if the key format is invalid.
+pub fn validate_btc_keypair(key: &str, role: &str) -> Result<Keypair> {
+    key.parse::<Keypair>()
+        .with_context(|| format!("Invalid Bitcoin private key for {role}"))
+}
+
+/// Parse and validate a Bitcoin public key.
+///
+/// Accepts both compressed (33 bytes) and uncompressed (65 bytes)
+/// public key formats in hex.
+///
+/// # Arguments
+///
+/// * `key` - The public key in hex format.
+/// * `role` - Description of the key's purpose (for error messages).
+///
+/// # Returns
+///
+/// The parsed public key.
+///
+/// # Errors
+///
+/// Returns an error if the key format is invalid.
+pub fn validate_btc_pubkey(key: &str, role: &str) -> Result<PublicKey> {
+    key.parse::<PublicKey>()
+        .with_context(|| format!("Invalid Bitcoin public key for {role}"))
+}
+
+/// Writes the generated secret and related data to a file with restricted permissions.
+///
+/// The file is created with mode 0600 (owner read/write only) to protect the secret.
+/// The format is a simple key-value text file for easy parsing.
+pub fn write_secret_to_file(
+    path: &Path,
+    secret: &[u8; 32],
+    secret_hash: &[u8; 32],
+    lock_txid: &Txid,
+) -> Result<()> {
+    if let Some(parent) = path.parent()
+        && !parent.as_os_str().is_empty()
+    {
+        fs::create_dir_all(parent)
+            .with_context(|| format!("Failed to create directory: {}", parent.display()))?;
+    }
+
+    // Create file with restricted permissions (Unix only)
+    let mut file = OpenOptions::new()
+        .write(true)
+        .create(true)
+        .truncate(true)
+        .mode(0o600)
+        .open(path)
+        .with_context(|| format!("Failed to create secret file: {}", path.display()))?;
+
+    writeln!(file, "# Atomic Swap Secret File")?;
+    writeln!(file, "SECRET={}", hex::encode(secret))?;
+    writeln!(file, "SECRET_HASH={}", hex::encode(secret_hash))?;
+    writeln!(file, "LOCK_TXID={lock_txid}")?;
+
+    Ok(())
+}
+
+/// Parses a secret file in the format generated by lock_bitcoin.
+///
+/// The file format is:
+/// ```text
+/// SECRET=<hex>
+/// SECRET_HASH=<hex>
+/// LOCK_TXID=<txid>
+/// ```
+///
+/// Lines starting with `#` are treated as comments and ignored.
+pub fn parse_secret_file(path: &Path) -> Result<SecretFileData> {
+    let content = std::fs::read_to_string(path)
+        .with_context(|| format!("Failed to read secret file: {}", path.display()))?;
+
+    let entries = content
+        .lines()
+        .map(str::trim)
+        .filter(|line| !line.is_empty() && !line.starts_with('#'))
+        .filter_map(|line| line.split_once('='))
+        .map(|(k, v)| (k.trim(), v.trim()))
+        .collect::<HashMap<&str, &str>>();
+
+    let secret = entries
+        .get("SECRET")
+        .ok_or_else(|| anyhow!("SECRET not found in file: {}", path.display()))
+        .and_then(|v| decode_hex_secret(v))?;
+
+    let secret_hash = entries
+        .get("SECRET_HASH")
+        .map(|v| decode_hex_hash(v, "secret hash"))
+        .transpose()?;
+
+    let lock_txid = entries
+        .get("LOCK_TXID")
+        .map(|v| v.parse().context("Invalid LOCK_TXID in secret file"))
+        .transpose()?;
+
+    Ok((secret, secret_hash, lock_txid))
+}
+
+/// Resolves a secret from either a direct hex string or a file.
+///
+/// # Errors
+///
+/// Returns an error if neither or both sources are provided.
+pub fn resolve_secret(secret: Option<String>, secret_file: Option<PathBuf>) -> Result<[u8; 32]> {
+    match (secret, secret_file) {
+        (Some(s), None) => decode_hex_secret(&s),
+        (None, Some(path)) => {
+            let (secret_bytes, _, _) = parse_secret_file(&path)?;
+            Ok(secret_bytes)
+        }
+        (None, None) => Err(anyhow!("Either --secret or --secret-file must be provided")),
+        (Some(_), Some(_)) => Err(anyhow!("--secret and --secret-file are mutually exclusive")),
+    }
+}
+
+/// Secret file data: (secret, optional secret_hash, optional lock_txid).
+type SecretFileData = ([u8; 32], Option<[u8; 32]>, Option<bitcoin::Txid>);
+
+/// Resolves a secret and optional metadata from either a direct hex string or a file.
+///
+/// When reading from a file, also extracts the secret hash and lock txid if present.
+///
+/// # Errors
+///
+/// Returns an error if neither or both sources are provided.
+pub fn resolve_secret_with_metadata(
+    secret: Option<String>,
+    secret_file: Option<PathBuf>,
+) -> Result<SecretFileData> {
+    match (secret, secret_file) {
+        (Some(s), None) => Ok((decode_hex_secret(&s)?, None, None)),
+        (None, Some(path)) => {
+            let (secret_bytes, secret_hash, lock_txid) = parse_secret_file(&path)?;
+            Ok((secret_bytes, secret_hash, lock_txid))
+        }
+        (None, None) => Err(anyhow!("Either --secret or --secret-file must be provided")),
+        (Some(_), Some(_)) => Err(anyhow!("--secret and --secret-file are mutually exclusive")),
+    }
+}
+
+/// Decodes a hex-encoded 32-byte hash value.
+///
+/// # Errors
+///
+/// Returns an error if the hex string is invalid or not exactly 32 bytes.
+pub fn decode_hex_hash(hex_str: &str, field_name: &str) -> Result<[u8; 32]> {
+    hex::decode(hex_str)
+        .with_context(|| format!("Invalid hex encoding for {field_name}"))
+        .and_then(|bytes| {
+            bytes.try_into().map_err(|v: Vec<u8>| {
+                anyhow!(
+                    "Invalid {field_name} length: expected 32 bytes, got {}",
+                    v.len()
+                )
+            })
+        })
+}
+
+/// Decodes a hex-encoded 32-byte secret.
+fn decode_hex_secret(hex_str: &str) -> Result<[u8; 32]> {
+    decode_hex_hash(hex_str, "secret")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn btc_network_parse() {
+        assert!(matches!(parse_btc_network("mainnet"), Ok(Network::Bitcoin)));
+        assert!(matches!(parse_btc_network("MAINNET"), Ok(Network::Bitcoin)));
+        assert!(matches!(parse_btc_network("testnet"), Ok(Network::Testnet)));
+        assert!(matches!(parse_btc_network("regtest"), Ok(Network::Regtest)));
+        assert!(matches!(parse_btc_network("signet"), Ok(Network::Signet)));
+        assert!(parse_btc_network("invalid").is_err());
+    }
+
+    #[test]
+    fn test_decode_hex_hash() {
+        let valid_hash = "a".repeat(64); // 32 bytes in hex
+        assert!(decode_hex_hash(&valid_hash, "test").is_ok());
+
+        let invalid_length = "a".repeat(30); // 15 bytes in hex
+        assert!(decode_hex_hash(&invalid_length, "test").is_err());
+
+        let invalid_hex = "zz".repeat(32);
+        assert!(decode_hex_hash(&invalid_hex, "test").is_err());
+    }
+}
